split to dedicated pr command
